Makefile: Run C preprocessor on dts files
Device tree files in kernel source use C defines for meaningful
names of some values. When device tree is compiled in kernel
source, the C preprocessor is run first on dts files and then
the dtc compiler.

To be able to use in overlays values with the same meaningful
values, run the C preprocessor first on the dts files. This
requires also the kernel source as the definitions of the
values are found there. When dtbo files are built it is now
required to set KERNEL_PATH variable with kernel source path.

Also use dtc from the kernel source path and remove
no-avoid_unnecessary_addr_size option as it is not avilable in
the dtc compiler from the latest lts kernel (4.14).

Signed-off-by: Todor Tomov <todor.tomov@linaro.org>